requirements.txt
flask==2.0.3
werkzeug==2.2.2
ultralytics==8.0.196
opencv-python-headless==4.5.5.64
google-cloud-storage==2.5.0
gunicorn==20.1.0
google-auth==2.17.3
google-auth-oauthlib==1.0.0
google-api-python-client==2.86.0
requests==2.28.1